pyproject.toml
[project]
name = "sftp-streaming"
version = "0.1.0"
description = "Streaming large files from SFTP to S3"
authors = [
    {name = "abyssnlp",email = "rawatshaurya1994@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "paramiko (>=3.0.0)",
    "boto3 (>=1.38.29,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3-stubs (>=1.38.29,<2.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "psutil (>=5.9.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.setuptools]

